requirements.txt
flask
mysql-connector-python
ultralytics
cvzone